add type to Map.DoubleClickZoom.ts
